refactor: credits whitelist to be at Notification [higher-level]
