+++ src/components/Notifications/NotificationGroupCard/NotificationGroupCard.types.ts
@@ -15,8 +15,6 @@ export type Props = {
   panelName: string
   onChangeNotificationSetting: typeof saveSubscriptionsRequest
   onChangeAccordion: (panel: string) => (event: React.SyntheticEvent, isExpanded: boolean) => void
-  whitelistedCreditsWallets?: string[]
-  address?: string
 }
 
 export type MapStateProps = Pick<
@@ -30,8 +28,6 @@ export type MapStateProps = Pick<
   | 'isExpanded'
   | 'panelName'
   | 'onChangeAccordion'
-  | 'whitelistedCreditsWallets'
-  | 'address'
 >
 export type OwnProps = Pick<
   Props,
