+++ src/components/Notifications/Notifications.container.ts
@@ -1,6 +1,7 @@
 import { connect } from 'react-redux'
 import { isLoadingType } from 'decentraland-dapps/dist/modules/loading/selectors'
-import { isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
+import { getAddress, isConnecting } from 'decentraland-dapps/dist/modules/wallet/selectors'
+import { getWhitelistedCreditsWallet } from '../../modules/features/selectors'
 import { RootState } from '../../modules/reducer'
 import { GET_SUBSCRIPTIONS_REQUEST, getSubscriptionsRequest } from '../../modules/subscription/actions'
 import { getLoading } from '../../modules/subscription/selectors'
@@ -9,7 +10,9 @@ import { MapDispatch, MapDispatchProps, MapStateProps, OwnProps } from './Notifi
 
 const mapState = (state: RootState, ownProps: OwnProps): MapStateProps => {
   return {
-    isLoading: ownProps.isLoading || isConnecting(state) || isLoadingType(getLoading(state), GET_SUBSCRIPTIONS_REQUEST)
+    isLoading: ownProps.isLoading || isConnecting(state) || isLoadingType(getLoading(state), GET_SUBSCRIPTIONS_REQUEST),
+    address: getAddress(state),
+    whitelistedCreditsWallets: getWhitelistedCreditsWallet(state)
   }
 }
